requirements.txt
PySide6>=6.5.0
python-docx>=0.8.11
openpyxl>=3.1.0
PyMuPDF>=1.23.0
python-pptx>=0.6.21
lxml>=4.9.0
chardet>=5.1.0
torch>=2.0.0
transformers>=4.30.0
huggingface_hub>=0.16.0
psutil>=5.9.0
